requirements.txt
streamlit
onnxruntime
numpy
pillow